stochastic rct with changes

+++ Digital Twin/dtwinpylib/dtwinpy/Digital_Twin.py
@@ -45,7 +45,7 @@ importlib.reload(dtwinpylib.dtwinpy.services)"""
 
 
 class Digital_Twin():
-    def __init__(self, name, copied_realDB= False,model_path= None, ip_address= None, initial= True, targeted_part_id= None, targeted_cluster= None, until= None, digital_database_path= None, real_database_path= None, ID_database_path= None, experimental_database_path= None, Freq_Sync= 1000, Freq_Valid= 10000, delta_t_treshold= 100, logic_threshold= 0.75, input_threshold= 0.75, rct_threshold= 0.02, queue_position= 2, Freq_Service = None, part_type= "A", loop_type= "closed", maxparts = None, template= False, keepDB= True, plot= False, verbose= True, flag_API= False, flag_external_service= False, flag_publish= True, flag_validation= False, rct_queue= 3):
+    def __init__(self, name, copied_realDB= False,model_path= None, ip_address= None, initial= True, targeted_part_id= None, targeted_cluster= None, until= None, digital_database_path= None, real_database_path= None, ID_database_path= None, experimental_database_path= None, Freq_Sync= 1000, Freq_Valid= 10000, delta_t_treshold= 100, logic_threshold= 0.75, input_threshold= 0.75, rct_threshold= 0.02, queue_position= 2, Freq_Service = None, part_type= "A", loop_type= "closed", maxparts = None, template= False, keepDB= True, keepModels= False, plot= False, verbose= True, flag_API= False, flag_external_service= False, flag_publish= True, flag_validation= False, rct_queue= 3):
         self.helper = Helper()
         #--- Model Parameters
         self.name = name
@@ -151,7 +151,8 @@ class Digital_Twin():
             self.model_path = f"{self.model_root}/{first_model}"
 
             #-- Clean existing model files
-            self.helper.delete_old_model(self.model_root, first_model)
+            if keepModels == False:
+                self.helper.delete_old_model(self.model_root, first_model)
         else:
             self.model_path = model_path
